// src/services/api/generated/inventory-service/model/seatResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Cinema Booking System API
 * API documentation for Cinema Booking System
 * OpenAPI spec version: 1.0.0
 */
import type { SeatResponseSeatType } from './seatResponseSeatType';

export interface SeatResponse {
  id?: string;
  roomId?: string;
  seatNumber?: string;
  rowLabel?: string;
  seatType?: SeatResponseSeatType;
  active?: boolean;
  createdAt?: string;
  updatedAt?: string;
}