// src/services/api/generated/inventory-service/model/showSeatResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Cinema Booking System API
 * API documentation for Cinema Booking System
 * OpenAPI spec version: 1.0.0
 */
import type { ShowSeatResponseSeatType } from './showSeatResponseSeatType';
import type { ShowSeatResponseStatus } from './showSeatResponseStatus';

export interface ShowSeatResponse {
  id?: string;
  showtimeId?: string;
  seatId?: string;
  seatNumber?: string;
  seatType?: ShowSeatResponseSeatType;
  price?: number;
  status?: ShowSeatResponseStatus;
  heldByBookingId?: string;
  holdExpiresAt?: string;
  createdAt?: string;
  updatedAt?: string;
}