// src/services/api/generated/inventory-service/model/showSeatResponseSeatType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Cinema Booking System API
 * API documentation for Cinema Booking System
 * OpenAPI spec version: 1.0.0
 */

export type ShowSeatResponseSeatType = typeof ShowSeatResponseSeatType[keyof typeof ShowSeatResponseSeatType];


export const ShowSeatResponseSeatType = {
  STANDARD: 'STANDARD',
  VIP: 'VIP',
  COUPLE: 'COUPLE',
  ACCESSIBLE: 'ACCESSIBLE',
} as const;